requirements.txt
asgiref==3.7.2
beautifulsoup4==4.12.3
blinker==1.7.0
bs4==0.0.2
cachetools==5.3.3
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
distlib==0.3.7
Django==4.2.7
django-chartit==0.2.9
django-chartjs==2.3.0
django-matplotlib==0.1
django-smart-selects==1.6.0
dpd-components==0.1.0
et-xmlfile==1.1.0
filelock==3.13.1
Flask==3.0.2
google-auth==2.28.2
google-auth-oauthlib==1.2.0
gspread==6.0.2
httplib2==0.22.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.3
lxml==5.1.0
MarkupSafe==2.1.3
mysql==0.0.3
mysqlclient==2.2.1
numpy==1.26.1
oauth2client==4.1.3
oauthlib==3.2.2
openpyxl==3.1.2
packaging==23.2
pandas==2.1.4
pdf2image==1.17.0
phonenumbers==8.13.24
pillow==10.2.0
platformdirs==3.11.0
plotly==5.18.0
pyasn1==0.5.1
pyasn1-modules==0.3.0
PyMySQL==1.1.0
pyparsing==3.1.2
pytesseract==0.3.10
python-dateutil==2.8.2
python-dotenv==1.0.1
python-pptx==0.6.23
pytz==2023.3.post1
requests==2.31.0
requests-oauthlib==1.4.0
rsa==4.9
six==1.16.0
soupsieve==2.5
sqlparse==0.4.4
StrEnum==0.4.15
tenacity==8.2.3
tesseract==0.1.3
tzdata==2023.3
urllib3==2.0.7
virtualenv==20.24.6
Werkzeug==3.0.1
XlsxWriter==3.2.0
